Created component for user actions
