+++ components/common/index.ts
@@ -15,3 +15,4 @@ export * from './StatusDot'
 export * from './SideDrawerCloseIcon'
 export * from './RangeSlider'
 export * from './ProgressionRing'
+export * from './ActionCall'
